Refactoring common libraries
