+++ common/src/test/java/com/google/udmi/util/GeneralUtilsTest.java
@@ -1,5 +1,12 @@
 package com.google.udmi.util;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertTrue;
+
+import com.google.common.collect.ImmutableMap;
+import java.util.HashMap;
+import java.util.Map;
 import org.junit.Test;
 
 /**
@@ -8,8 +15,15 @@ import org.junit.Test;
 public class GeneralUtilsTest {
 
   @Test
+  @SuppressWarnings("unchecked")
   public void deepCopy() {
-    throw new RuntimeException("Nope");
+    Map<String, Object> original = new HashMap<>();
+    original.put("A", "B");
+    original.put("C", ImmutableMap.of("D", "E"));
+    Map<String, Object> copy = GeneralUtils.deepCopy(original);
+    assertEquals("copy is equal", original, copy);
+    ((Map<String, String>) copy.get("C")).put("F", "G");
+    assertNotEquals("copy is not equal", original, copy);
   }
 
   @Test
