+++ validator/src/main/java/com/google/daq/mqtt/registrar/LocalDevice.java
@@ -7,7 +7,7 @@ import static com.google.daq.mqtt.registrar.Registrar.ENVELOPE_JSON;
 import static com.google.daq.mqtt.registrar.Registrar.GENERATED_CONFIG_JSON;
 import static com.google.daq.mqtt.registrar.Registrar.METADATA_JSON;
 import static com.google.daq.mqtt.registrar.Registrar.NORMALIZED_JSON;
-import static com.google.daq.mqtt.util.JsonUtil.asMap;
+import static com.google.udmi.util.JsonUtil.asMap;
 
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.core.JsonGenerator;
@@ -34,10 +34,10 @@ import com.google.daq.mqtt.util.CloudDeviceSettings;
 import com.google.daq.mqtt.util.CloudIotManager;
 import com.google.daq.mqtt.util.ExceptionMap;
 import com.google.daq.mqtt.util.ExceptionMap.ErrorTree;
-import com.google.daq.mqtt.util.JsonUtil;
 import com.google.daq.mqtt.util.MessageUpgrader;
 import com.google.daq.mqtt.util.ValidationException;
 import com.google.udmi.util.GeneralUtils;
+import com.google.udmi.util.JsonUtil;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
