+++ validator/src/main/java/com/google/daq/mqtt/sequencer/sequences/DiscoverySequences.java
@@ -8,8 +8,8 @@ import static org.junit.Assert.fail;
 import com.google.daq.mqtt.sequencer.SequenceBase;
 import com.google.daq.mqtt.sequencer.SkipTest;
 import com.google.daq.mqtt.sequencer.semantic.SemanticDate;
-import com.google.daq.mqtt.util.JsonUtil;
-import com.google.daq.mqtt.validator.CleanDateFormat;
+import com.google.udmi.util.CleanDateFormat;
+import com.google.udmi.util.JsonUtil;
 import java.time.Instant;
 import java.util.Date;
 import java.util.HashMap;
