+++ validator/src/main/java/com/google/daq/mqtt/sequencer/sequences/WritebackSequences.java
@@ -1,7 +1,7 @@
 package com.google.daq.mqtt.sequencer.sequences;
 
 import com.google.daq.mqtt.sequencer.PointSequencer;
-import com.google.daq.mqtt.util.JsonUtil;
+import com.google.udmi.util.JsonUtil;
 import java.util.Objects;
 import org.junit.Test;
 import udmi.schema.PointPointsetState.Value_state;
