+++ validator/src/main/java/com/google/daq/mqtt/validator/Validator.java
@@ -8,8 +8,8 @@ import static com.google.daq.mqtt.util.Common.TIMESTAMP_ATTRIBUTE;
 import static com.google.daq.mqtt.util.Common.removeNextArg;
 import static com.google.daq.mqtt.util.ConfigUtil.UDMI_VERSION;
 import static com.google.daq.mqtt.util.ConfigUtil.readExecutionConfiguration;
-import static com.google.daq.mqtt.util.JsonUtil.JSON_SUFFIX;
-import static com.google.daq.mqtt.util.JsonUtil.OBJECT_MAPPER;
+import static com.google.udmi.util.JsonUtil.JSON_SUFFIX;
+import static com.google.udmi.util.JsonUtil.OBJECT_MAPPER;
 import static java.util.Objects.requireNonNull;
 
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -31,12 +31,12 @@ import com.google.daq.mqtt.util.ConfigUtil;
 import com.google.daq.mqtt.util.ExceptionMap;
 import com.google.daq.mqtt.util.ExceptionMap.ErrorTree;
 import com.google.daq.mqtt.util.FileDataSink;
-import com.google.daq.mqtt.util.JsonUtil;
 import com.google.daq.mqtt.util.MessagePublisher;
 import com.google.daq.mqtt.util.MessageUpgrader;
 import com.google.daq.mqtt.util.PubSubClient;
 import com.google.daq.mqtt.util.ValidationException;
 import com.google.udmi.util.GeneralUtils;
+import com.google.udmi.util.JsonUtil;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
